release v1.0.0

+++ ddth-dlock-core/src/test/java/com/github/ddth/dlock/test/clusteredredis/ClusteredRedisDLockTCase.java
@@ -1,16 +1,14 @@
 package com.github.ddth.dlock.test.clusteredredis;
 
+import com.github.ddth.dlock.impl.redis.ClusteredRedisDLockFactory;
+import com.github.ddth.dlock.test.BaseDLockTCase;
+import org.junit.Before;
+import redis.clients.jedis.Jedis;
+import redis.clients.jedis.JedisCluster;
+
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Properties;
-
-import org.junit.Before;
-
-import com.github.ddth.dlock.impl.redis.ClusteredRedisDLockFactory;
-import com.github.ddth.dlock.test.BaseDLockTCase;
-
-import redis.clients.jedis.Jedis;
-import redis.clients.jedis.JedisCluster;
 
 /**
  * @author Thanh Nguyen <btnguyen2k@gmail.com>
@@ -19,13 +17,24 @@ import redis.clients.jedis.JedisCluster;
 public class ClusteredRedisDLockTCase extends BaseDLockTCase {
 
     protected static class MyClusteredRedisDLockFactory extends ClusteredRedisDLockFactory {
+        private boolean isMaster(String name, Jedis node) {
+            for (String nodeInfo : node.clusterNodes().split("\n")) {
+                if (nodeInfo.contains(name)) {
+                    return nodeInfo.contains("master");
+                }
+            }
+            return false;
+        }
+
         public MyClusteredRedisDLockFactory init() {
             super.init();
 
             JedisCluster jedis = getJedisConnector().getJedisCluster();
             jedis.getClusterNodes().forEach((name, pool) -> {
                 try (Jedis node = pool.getResource()) {
-                    node.flushAll();
+                    if (isMaster(name, node)) {
+                        node.flushAll();
+                    }
                 }
             });
 
@@ -38,7 +47,8 @@ public class ClusteredRedisDLockTCase extends BaseDLockTCase {
         Map<String, Properties> lockProperties = new HashMap<>();
 
         MyClusteredRedisDLockFactory factory = new MyClusteredRedisDLockFactory();
-        factory.setRedisHostsAndPorts("localhost:7000");
+        factory.setRedisHostsAndPorts(
+                "localhost:7000,localhost:7001,localhost:7002,localhost:7003,localhost:7004,localhost:7005");
         factory.setLockProperties(lockProperties);
         factory.init();
         lockFactory = factory;
